src/lib.rs: work against more Slurm versions
In version 17.11, the type `List` is a `struct xlist *`, but in 17.02 it is a
`struct list *`. We only ever access these structs as pointers, so by using
the `slurm_sys::List` type we can work with both interfaces. I'm not a fan of
the type hiding the fact that it's a pointer, but that's pretty small potatoes
overall.