Implement a system for portability across C API revisions.
This feels pretty gross but in the end I think it might be a pretty decent
approach. After calling bindgen, we scan the generated Rust code and check for
various features that differ between the two Slurm versions I'm investigating
(17.02 and 17.11). We augment the generated `slurm_sys` crate with a constant
item reporting what we've found. The main `slurm` crate then acquires a new
`build.rs` script that looks at this magic constant and translates into
feature flags that are passed to rustc (via printing `cargo:rust-cfg=...`
lines).

This gives us the tools we need to do some conditional compilation in the main
module, without exposing the guts of all our weird little features externally.