More work on docs.rs building
I recently discovered that the "dependencies = [debian-packages...]" feature
isn't actually implemented, so I'll need to do something else to get slurm-sys
compiling. As best I can see, I might need to include a canned version of the
clang-generated Rust bindings, plus not try to link with the libraries (which
is OK since for docs we don't need to actually run any code).